backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Testing==0.8.1 
python-dotenv==1.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.17.1
pytz==2023.3
reportlab==4.4.4
openpyxl==3.1.2
coverage==7.3.2